Pull registry implementation out of RMI Remote Object Config.
Pull the remote instance registry implementation out of the RMI Remote
Object Config, and introduce an interface for getting and registering
method argument types. This cleans up the remote object config a
little bit, as by convention there are two instance registry
implementations: the actual registry instantiated by the remote object
config, and the switching registry.

Previously the RMI Remote Object Config class had too much knowledge
of the internals of the registry. In the future, for testing purposes,
we should be able to provide an alternative implementation for the
registry (e.g. to assert calls made by the RMI Remote Object against
it).
